Update gorunov1.cpp
Added optimizations. Global namespace was std so there was no need to have the scope resolution of std:: in the code. Removed return 0 because in modern C++ you are not required to do it anymore because it happens by default. Added auto keyword in for loop for modern semantics. It deduces the appropriate datatype on its own.